# spiffworkflow-backend/src/spiffworkflow_backend/api.yml
openapi: "3.0.2"
info:
  version: 1.0.0
  title: spiffworkflow-backend
  license:
    name: LGPL
servers:
  - url: http://localhost:5000/v1.0
# this is handled in flask now
security: []
paths:
...
  /upsearch-locations:
    parameters:
      - name: location
        in: query
        required: false
        description: The location to perform the upsearch
        schema:
          type: string
    get:
      operationId: spiffworkflow_backend.routes.upsearch_controller.upsearch_locations
      summary: Returns upsearch locations for the given location
      tags:
        - Upsearch
      responses:
        "200":
          description: Upsearch Locations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Locations"
...
  /message-models/{relative_location}:
    parameters:
      - name: relative_location
        in: path
        required: true
        description: The location of the message model relative to the root of the project, defaults to /
        schema:
          type: string
    get:
      tags:
        - Messages
      operationId: spiffworkflow_backend.routes.messages_controller.message_model_list
      summary: Get a list of message models at the specified location
      responses:
        "200":
          description: A list of message models with their correlation properties at the specified location and higher in the directory tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageModelListResponse"
  /messages:
    parameters:
      - name: process_instance_id
        in: query
        required: false
        description: the id of the process instance
        schema:
          type: integer
      - name: page
        in: query
        required: false
        description: The page number to return. Defaults to page 1.
        schema:
          type: integer
      - name: per_page
        in: query
        required: false
        description: The number of models to show per page. Defaults to 10.
        schema:
          type: integer
    get:
      tags:
        - Messages
      operationId: spiffworkflow_backend.routes.messages_controller.message_instance_list
      summary: Get a list of message instances
      responses:
        "200":
          description: One task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AwesomeUnspecifiedPayload"

    post:
      operationId: spiffworkflow_backend.routes.messages_controller.message_instance_search
      summary: Returns a list of message instances matching the search criteria
      tags:
        - Messages
      requestBody:
        description: Filtering criteria for search
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                process_model_identifier:
                  type: string
                  nullable: true
                  description: Process Model
                name:
                  type: string
                  nullable: true
                  description: Message name identifier
                message_type:
                  type: array
                  items:
                    type: string
                  nullable: true
                  description: Send or receive message
                status:
                  type: array
                  items:
                    type: string
                  nullable: true
                  description: Message status
                created_after:
                  type: integer
                  nullable: true
                  description: Messages sent or received after timestamp
                created_before:
                  type: integer
                  nullable: true
                  description: Messages sent or received before timestamp
      responses:
        "200":
          description: A filtered list of messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AwesomeUnspecifiedPayload"
  /messages/{modified_message_name}:
    parameters:
      - name: modified_message_name
        in: path
        required: true
        description: The message_name, modified to replace slashes (/) with colons
        schema:
          type: string
      - name: execution_mode
        in: query
        required: false
        description: Either run in "synchronous" or "asynchronous" mode.
        schema:
          type: string
          enum:
            - synchronous
            - asynchronous
      - name: time_to_live_in_seconds
        in: query
        required: false
        description: Seconds to buffer an unmatched message. A value of 0 means the message must match immediately; rejected messages may be retained with status not_accepted for audit and debugging.
        schema:
          type: integer
          minimum: 0
          maximum: 300
      - name: message_instance_uuid
        in: query
        required: false
        description: Caller-provided idempotency key. Required when time_to_live_in_seconds is greater than 0.
        schema:
          type: string
          format: uuid
          maxLength: 40
          pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
    post:
      tags:
        - Messages
      operationId: spiffworkflow_backend.routes.messages_controller.message_send
      summary: Instantiate and run a given process model with a message start event matching given name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AwesomeUnspecifiedPayload"
      responses:
        "200":
          description: One task
          content:
            application/json:
              schema:
                properties:
                  task_data:
                    $ref: "#/components/schemas/AwesomeUnspecifiedPayload"
                  process_instance:
                    $ref: "#/components/schemas/AwesomeUnspecifiedPayload"
        "202":
          description: Message accepted and buffered until a matching receiver is available
          content:
            application/json:
              schema:
                properties:
                  message_instance:
                    $ref: "#/components/schemas/AwesomeUnspecifiedPayload"
  /messages/{modified_message_name}/process-model:
    parameters:
      - name: modified_message_name
        in: path
        required: true
        description: The message_name, modified to replace slashes (/) with colons
        schema:
          type: string
    get:
      operationId: spiffworkflow_backend.routes.messages_controller.get_process_model_for_message
      summary: Returns the process model for a given message name
      tags:
        - Messages
      responses:
        "200":
          description: The process model associated with the message.
          content:
            application/json:
              schema:
                type: object
                properties:
                  process_model_identifier:
                    type: string
                  file_name:
                    type: string
...
components:
  schemas:
...
    MessageCorrelationProperty:
      type: object
      properties:
        identifier:
          type: string
          description: Unique identifier for the correlation property
        retrieval_expression:
          type: string
          description: Expression used to retrieve the correlation value
      required:
        - identifier
        - retrieval_expression
    MessageModel:
      type: object
      properties:
        identifier:
          type: string
          description: Unique identifier for the message model
        location:
          type: string
          description: Location path where the message model is defined
        schema:
          type: object
          description: JSON schema defining the structure of the message payload
        correlation_properties:
          type: array
          items:
            $ref: "#/components/schemas/MessageCorrelationProperty"
          description: List of correlation properties used for message correlation
      required:
        - identifier
        - location
        - schema
        - correlation_properties
    MessageModelListResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/MessageModel"
          description: List of message models
      required:
        - messages
...
    Locations:
      properties:
        locations:
          type: array
...
    AwesomeUnspecifiedPayload:
      type: "object"
      additionalProperties: {}
...